rename utils file's name to custom hooks name
